reduce dependencies of llb package
